change return type of estimator, use double instead of int. be consistent with timer.

+++ src/estimator.h
@@ -7,19 +7,19 @@
 *                   MPI Functions for Management                  *
 *                                                                 *
 ******************************************************************/
-int  E_MPI_Init(
+double E_MPI_Init(
 int * argc,
 char *** argv );
 
 #ifdef PERF_MPI_THREADED
-int  E_MPI_Init_thread (
+double E_MPI_Init_thread (
 int * argc,
 char *** argv,
 int required,
 int *provided );
 
 #endif /* PERF_MPI_THREADED */
-int  E_MPI_Finalize();
+double E_MPI_Finalize();
 
 
 /******************************************************************
@@ -27,7 +27,7 @@ int  E_MPI_Finalize();
 *          MPI Point-to-Point(
 *                                                                 *
 ******************************************************************/
-int  E_MPI_Bsend(
+double E_MPI_Bsend(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -35,7 +35,7 @@ int dest,
 int tag,
 MPI_Comm comm );
 
-int  E_MPI_Bsend_init(
+double E_MPI_Bsend_init(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -44,7 +44,7 @@ int tag,
 MPI_Comm comm,
 MPI_Request * request );
 
-int  E_MPI_Recv_init(
+double E_MPI_Recv_init(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -53,7 +53,7 @@ int tag,
 MPI_Comm comm,
 MPI_Request * request );
 
-int  E_MPI_Send_init(
+double E_MPI_Send_init(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -62,7 +62,7 @@ int tag,
 MPI_Comm comm,
 MPI_Request * request );
 
-int  E_MPI_Ibsend(
+double E_MPI_Ibsend(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -70,7 +70,7 @@ int dest,
 int tag,
 MPI_Comm comm,
 MPI_Request * request );
-int  E_MPI_Irecv(
+double E_MPI_Irecv(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -78,7 +78,7 @@ int source,
 int tag,
 MPI_Comm comm,
 MPI_Request * request );
-int  E_MPI_Irsend(
+double E_MPI_Irsend(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -86,7 +86,7 @@ int dest,
 int tag,
 MPI_Comm comm,
 MPI_Request * request );
-int  E_MPI_Isend(
+double E_MPI_Isend(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -94,7 +94,7 @@ int dest,
 int tag,
 MPI_Comm comm,
 MPI_Request * request );
-int  E_MPI_Issend(
+double E_MPI_Issend(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -102,7 +102,7 @@ int dest,
 int tag,
 MPI_Comm comm,
 MPI_Request * request );
-int  E_MPI_Recv(
+double E_MPI_Recv(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -110,14 +110,14 @@ int source,
 int tag,
 MPI_Comm comm,
 MPI_Status * status );
-int  E_MPI_Rsend(
+double E_MPI_Rsend(
 void * buf,
 int count,
 MPI_Datatype datatype,
 int dest,
 int tag,
 MPI_Comm comm );
-int  E_MPI_Rsend_init(
+double E_MPI_Rsend_init(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -125,14 +125,14 @@ int dest,
 int tag,
 MPI_Comm comm,
 MPI_Request * request );
-int  E_MPI_Send(
+double E_MPI_Send(
 void * buf,
 int count,
 MPI_Datatype datatype,
 int dest,
 int tag,
 MPI_Comm comm );
-int  E_MPI_Sendrecv(
+double E_MPI_Sendrecv(
 void * sendbuf,
 int sendcount,
 MPI_Datatype sendtype,
@@ -145,7 +145,7 @@ int source,
 int recvtag,
 MPI_Comm comm,
 MPI_Status * status );
-int  E_MPI_Sendrecv_replace(
+double E_MPI_Sendrecv_replace(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -155,14 +155,14 @@ int source,
 int recvtag,
 MPI_Comm comm,
 MPI_Status * status );
-int  E_MPI_Ssend(
+double E_MPI_Ssend(
 void * buf,
 int count,
 MPI_Datatype datatype,
 int dest,
 int tag,
 MPI_Comm comm );
-int  E_MPI_Ssend_init(
+double E_MPI_Ssend_init(
 void * buf,
 int count,
 MPI_Datatype datatype,
@@ -170,64 +170,64 @@ int dest,
 int tag,
 MPI_Comm comm,
 MPI_Request * request );
-int   E_MPI_Test(
+double E_MPI_Test(
 MPI_Request * request,
 int * flag,
 MPI_Status * status );
-int  E_MPI_Testall(
+double E_MPI_Testall(
 int count,
 MPI_Request * array_of_requests,
 int * flag,
 MPI_Status * array_of_statuses );
-int  E_MPI_Testany(
+double E_MPI_Testany(
 int count,
 MPI_Request * array_of_requests,
 int * index,
 int * flag,
 MPI_Status * status );
-int  E_MPI_Test_cancelled(
+double E_MPI_Test_cancelled(
 MPI_Status * status,
 int * flag );
-int  E_MPI_Testsome(
+double E_MPI_Testsome(
 int incount,
 MPI_Request * array_of_requests,
 int * outcount,
 int * array_of_indices,
 MPI_Status * array_of_statuses );
-int   E_MPI_Wait(
+double E_MPI_Wait(
 MPI_Request * request,
 MPI_Status * status );
-int  E_MPI_Waitall(
+double E_MPI_Waitall(
 int count,
 MPI_Request * array_of_requests,
 MPI_Status * array_of_statuses );
-int  E_MPI_Waitany(
+double E_MPI_Waitany(
 int count,
 MPI_Request * array_of_requests,
 int * index,
 MPI_Status * status );
-int  E_MPI_Waitsome(
+double E_MPI_Waitsome(
 int incount,
 MPI_Request * array_of_requests,
 int * outcount,
 int * array_of_indices,
 MPI_Status * array_of_statuses );
-int  E_MPI_Cancel(
+double E_MPI_Cancel(
 MPI_Request * request );
-int  E_MPI_Request_free(
+double E_MPI_Request_free(
 MPI_Request * request );
-int  E_MPI_Start(
+double E_MPI_Start(
 MPI_Request * request );
-int  E_MPI_Startall(
+double E_MPI_Startall(
 int count,
 MPI_Request * array_of_requests );
-int  E_MPI_Iprobe(
+double E_MPI_Iprobe(
 int source,
 int tag,
 MPI_Comm comm,
 int * flag,
 MPI_Status * status );
-int  E_MPI_Probe(
+double E_MPI_Probe(
 int source,
 int tag,
 MPI_Comm comm,
@@ -237,7 +237,7 @@ MPI_Status * status );
 *               MPI Collective Communication Related              *
 *                                                                 *
 ******************************************************************/
-int   E_MPI_Allgather(
+double E_MPI_Allgather(
 void * sendbuf,
 int sendcount,
 MPI_Datatype sendtype,
@@ -245,7 +245,7 @@ void * recvbuf,
 int recvcount,
 MPI_Datatype recvtype,
 MPI_Comm comm );
-int   E_MPI_Allgatherv(
+double E_MPI_Allgatherv(
 void * sendbuf,
 int sendcount,
 MPI_Datatype sendtype,
@@ -254,14 +254,14 @@ int * recvcounts,
 int * displs,
 MPI_Datatype recvtype,
 MPI_Comm comm );
-int   E_MPI_Allreduce(
+double E_MPI_Allreduce(
 void * sendbuf,
 void * recvbuf,
 int count,
 MPI_Datatype datatype,
 MPI_Op op,
 MPI_Comm comm );
-int  E_MPI_Alltoall(
+double E_MPI_Alltoall(
 void * sendbuf,
 int sendcount,
 MPI_Datatype sendtype,
@@ -269,7 +269,7 @@ void * recvbuf,
 int recvcnt,
 MPI_Datatype recvtype,
 MPI_Comm comm );
-int   E_MPI_Alltoallv(
+double E_MPI_Alltoallv(
 void * sendbuf,
 int * sendcnts,
 int * sdispls,
@@ -279,15 +279,15 @@ int * recvcnts,
 int * rdispls,
 MPI_Datatype recvtype,
 MPI_Comm comm );
-int   E_MPI_Barrier(
+double E_MPI_Barrier(
 MPI_Comm comm );
-int   E_MPI_Bcast(
+double E_MPI_Bcast(
 void * buffer,
 int count,
 MPI_Datatype datatype,
 int root,
 MPI_Comm comm );
-int   E_MPI_Gather(
+double E_MPI_Gather(
 void * sendbuf,
 int sendcnt,
 MPI_Datatype sendtype,
@@ -296,7 +296,7 @@ int recvcount,
 MPI_Datatype recvtype,
 int root,
 MPI_Comm comm );
-int   E_MPI_Gatherv(
+double E_MPI_Gatherv(
 void * sendbuf,
 int sendcnt,
 MPI_Datatype sendtype,
@@ -306,14 +306,14 @@ int * displs,
 MPI_Datatype recvtype,
 int root,
 MPI_Comm comm );
-int   E_MPI_Reduce_scatter(
+double E_MPI_Reduce_scatter(
 void * sendbuf,
 void * recvbuf,
 int * recvcnts,
 MPI_Datatype datatype,
 MPI_Op op,
 MPI_Comm comm );
-int   E_MPI_Reduce(
+double E_MPI_Reduce(
 void * sendbuf,
 void * recvbuf,
 int count,
@@ -321,14 +321,14 @@ MPI_Datatype datatype,
 MPI_Op op,
 int root,
 MPI_Comm comm );
-int   E_MPI_Scan(
+double E_MPI_Scan(
 void * sendbuf,
 void * recvbuf,
 int count,
 MPI_Datatype datatype,
 MPI_Op op,
 MPI_Comm comm );
-int   E_MPI_Scatter(
+double E_MPI_Scatter(
 void * sendbuf,
 int sendcnt,
 MPI_Datatype sendtype,
@@ -337,7 +337,7 @@ int recvcnt,
 MPI_Datatype recvtype,
 int root,
 MPI_Comm comm );
-int   E_MPI_Scatterv(
+double E_MPI_Scatterv(
 void * sendbuf,
 int * sendcnts,
 int * displs,
@@ -352,188 +352,188 @@ MPI_Comm comm );
 *               MPI Communicator Related Functions                *
 *                                                                 *
 ******************************************************************/
-int   E_MPI_Comm_compare(
+double E_MPI_Comm_compare(
 MPI_Comm comm1,
 MPI_Comm comm2,
 int * result );
-int   E_MPI_Comm_create(
+double E_MPI_Comm_create(
 MPI_Comm comm,
 MPI_Group group,
 MPI_Comm * comm_out );
-int   E_MPI_Comm_dup(
+double E_MPI_Comm_dup(
 MPI_Comm comm,
 MPI_Comm * comm_out );
-int   E_MPI_Comm_free(
+double E_MPI_Comm_free(
 MPI_Comm * comm );
-int   E_MPI_Comm_group(
+double E_MPI_Comm_group(
 MPI_Comm comm,
 MPI_Group * group );
-int   E_MPI_Comm_rank(
+double E_MPI_Comm_rank(
 MPI_Comm comm,
 int * rank );
-int   E_MPI_Comm_remote_group(
+double E_MPI_Comm_remote_group(
 MPI_Comm comm,
 MPI_Group * group );
-int   E_MPI_Comm_remote_size(
+double E_MPI_Comm_remote_size(
 MPI_Comm comm,
 int * size );
-int   E_MPI_Comm_size(
+double E_MPI_Comm_size(
 MPI_Comm comm,
 int * size );
-int   E_MPI_Comm_split(
+double E_MPI_Comm_split(
 MPI_Comm comm,
 int color,
 int key,
 MPI_Comm * comm_out );
-int   E_MPI_Comm_test_inter(
+double E_MPI_Comm_test_inter(
 MPI_Comm comm,
 int * flag );
-int   E_MPI_Group_compare(
+double E_MPI_Group_compare(
 MPI_Group group1,
 MPI_Group group2,
 int * result );
-int   E_MPI_Group_difference(
+double E_MPI_Group_difference(
 MPI_Group group1,
 MPI_Group group2,
 MPI_Group * group_out );
-int   E_MPI_Group_excl(
+double E_MPI_Group_excl(
 MPI_Group group,
 int n,
 int * ranks,
 MPI_Group * newgroup );
-int   E_MPI_Group_free(
+double E_MPI_Group_free(
 MPI_Group * group );
-int   E_MPI_Group_incl(
+double E_MPI_Group_incl(
 MPI_Group group,
 int n,
 int * ranks,
 MPI_Group * group_out );
-int   E_MPI_Group_intersection(
+double E_MPI_Group_intersection(
 MPI_Group group1,
 MPI_Group group2,
 MPI_Group * group_out );
-int   E_MPI_Group_rank(
+double E_MPI_Group_rank(
 MPI_Group group,
 int * rank );
-int   E_MPI_Group_range_excl(
+double E_MPI_Group_range_excl(
 MPI_Group group,
 int n,
 int ranges[][3],
 MPI_Group * newgroup );
-int   E_MPI_Group_range_incl(
+double E_MPI_Group_range_incl(
 MPI_Group group,
 int n,
 int ranges[][3],
 MPI_Group * newgroup );
-int   E_MPI_Group_size(
+double E_MPI_Group_size(
 MPI_Group group,
 int * size );
-int   E_MPI_Group_translate_ranks(
+double E_MPI_Group_translate_ranks(
 MPI_Group group_a,
 int n,
 int * ranks_a,
 MPI_Group group_b,
 int * ranks_b );
-int   E_MPI_Group_union(
+double E_MPI_Group_union(
 MPI_Group group1,
 MPI_Group group2,
 MPI_Group * group_out );
-int   E_MPI_Intercomm_create(
+double E_MPI_Intercomm_create(
 MPI_Comm local_comm,
 int local_leader,
 MPI_Comm peer_comm,
 int remote_leader,
 int tag,
 MPI_Comm * comm_out );
-int   E_MPI_Intercomm_merge(
+double E_MPI_Intercomm_merge(
 MPI_Comm comm,
 int high,
 MPI_Comm * comm_out );
-int   E_MPI_Keyval_create(
+double E_MPI_Keyval_create(
 MPI_Copy_function * copy_fn,
 MPI_Delete_function * delete_fn,
 int * keyval,
 void * extra_state );
-int   E_MPI_Keyval_free(
+double E_MPI_Keyval_free(
 int * keyval );
-int   E_MPI_Cart_coords(
+double E_MPI_Cart_coords(
 MPI_Comm comm,
 int rank,
 int maxdims,
 int * coords );
-int   E_MPI_Cart_create(
+double E_MPI_Cart_create(
 MPI_Comm comm_old,
 int ndims,
 int * dims,
 int * periods,
 int reorder,
 MPI_Comm * comm_cart );
-int   E_MPI_Cart_get(
+double E_MPI_Cart_get(
 MPI_Comm comm,
 int maxdims,
 int * dims,
 int * periods,
 int * coords );
-int   E_MPI_Cart_map(
+double E_MPI_Cart_map(
 MPI_Comm comm_old,
 int ndims,
 int * dims,
 int * periods,
 int * newrank );
-int   E_MPI_Cart_rank(
+double E_MPI_Cart_rank(
 MPI_Comm comm,
 int * coords,
 int * rank );
-int   E_MPI_Cart_shift(
+double E_MPI_Cart_shift(
 MPI_Comm comm,
 int direction,
 int displ,
 int * source,
 int * dest );
-int   E_MPI_Cart_sub(
+double E_MPI_Cart_sub(
 MPI_Comm comm,
 int * remain_dims,
 MPI_Comm * comm_new );
-int   E_MPI_Cartdim_get(
+double E_MPI_Cartdim_get(
 MPI_Comm comm,
 int * ndims );
-int  E_MPI_Dims_create(
+double E_MPI_Dims_create(
 int nnodes,
 int ndims,
 int * dims );
-int   E_MPI_Graph_create(
+double E_MPI_Graph_create(
 MPI_Comm comm_old,
 int nnodes,
 int * index,
 int * edges,
 int reorder,
 MPI_Comm * comm_graph );
-int   E_MPI_Graph_get(
+double E_MPI_Graph_get(
 MPI_Comm comm,
 int maxindex,
 int maxedges,
 int * index,
 int * edges );
-int   E_MPI_Graph_map(
+double E_MPI_Graph_map(
 MPI_Comm comm_old,
 int nnodes,
 int * index,
 int * edges,
 int * newrank );
-int   E_MPI_Graph_neighbors(
+double E_MPI_Graph_neighbors(
 MPI_Comm comm,
 int rank,
 int  maxneighbors,
 int * neighbors );
-int   E_MPI_Graph_neighbors_count(
+double E_MPI_Graph_neighbors_count(
 MPI_Comm comm,
 int rank,
 int * nneighbors );
-int   E_MPI_Graphdims_get(
+double E_MPI_Graphdims_get(
 MPI_Comm comm,
 int * nnodes,
 int * nedges );
-int   E_MPI_Topo_test(
+double E_MPI_Topo_test(
 MPI_Comm comm,
 int * top_type );
 /******************************************************************
@@ -548,115 +548,115 @@ int  MPI_Abort( MPI_Comm comm, int errorcode );
 #endif /* MPI_Abort & LAM MPI [LAM MPI] */
 
 
-int  E_MPI_Error_class(
+double E_MPI_Error_class(
 int errorcode,
 int * errorclass );
-int  E_MPI_Errhandler_create(
+double E_MPI_Errhandler_create(
 MPI_Handler_function * function,
 MPI_Errhandler * errhandler );
-int  E_MPI_Errhandler_free(
+double E_MPI_Errhandler_free(
 MPI_Errhandler * errhandler );
-int  E_MPI_Errhandler_get(
+double E_MPI_Errhandler_get(
 MPI_Comm comm,
 MPI_Errhandler * errhandler );
-int  E_MPI_Error_string(
+double E_MPI_Error_string(
 int errorcode,
 char * string,
 int * resultlen );
-int  E_MPI_Errhandler_set(
+double E_MPI_Errhandler_set(
 MPI_Comm comm,
 MPI_Errhandler errhandler );
-int  E_MPI_Get_processor_name(
+double E_MPI_Get_processor_name(
 char * name,
 int * resultlen );
 double  E_MPI_Wtick();
 double  E_MPI_Wtime();
-int  E_MPI_Address(
+double E_MPI_Address(
 void * location,
 MPI_Aint * address );
-int  E_MPI_Op_create(
+double E_MPI_Op_create(
 MPI_User_function * function,
 int commute,
 MPI_Op * op );
-int  E_MPI_Op_free(
+double E_MPI_Op_free(
 MPI_Op * op );
-int   E_MPI_Attr_delete(
+double E_MPI_Attr_delete(
 MPI_Comm comm,
 int keyval );
-int   E_MPI_Attr_get(
+double E_MPI_Attr_get(
 MPI_Comm comm,
 int keyval,
 void * attr_value,
 int * flag );
-int   E_MPI_Attr_put(
+double E_MPI_Attr_put(
 MPI_Comm comm,
 int keyval,
 void * attr_value );
-int  E_MPI_Buffer_attach(
+double E_MPI_Buffer_attach(
 void * buffer,
 int size );
-int  E_MPI_Buffer_detach(
+double E_MPI_Buffer_detach(
 void * buffer,
 int * size );
-int   E_MPI_Get_elements(
+double E_MPI_Get_elements(
 MPI_Status * status,
 MPI_Datatype datatype,
 int * elements );
-int  E_MPI_Get_count(
+double E_MPI_Get_count(
 MPI_Status * status,
 MPI_Datatype datatype,
 int * count );
-int   E_MPI_Type_commit(
+double E_MPI_Type_commit(
 MPI_Datatype * datatype );
-int  E_MPI_Type_contiguous(
+double E_MPI_Type_contiguous(
 int count,
 MPI_Datatype old_type,
 MPI_Datatype * newtype );
-int  E_MPI_Type_extent(
+double E_MPI_Type_extent(
 MPI_Datatype datatype,
 MPI_Aint * extent );
-int   E_MPI_Type_free(
+double E_MPI_Type_free(
 MPI_Datatype * datatype );
-int  E_MPI_Type_hindexed(
+double E_MPI_Type_hindexed(
 int count,
 int * blocklens,
 MPI_Aint * indices,
 MPI_Datatype old_type,
 MPI_Datatype * newtype );
-int  E_MPI_Type_hvector(
+double E_MPI_Type_hvector(
 int count,
 int blocklen,
 MPI_Aint stride,
 MPI_Datatype old_type,
 MPI_Datatype * newtype );
-int  E_MPI_Type_indexed(
+double E_MPI_Type_indexed(
 int count,
 int * blocklens,
 int * indices,
 MPI_Datatype old_type,
 MPI_Datatype * newtype );
-int   E_MPI_Type_lb(
+double E_MPI_Type_lb(
 MPI_Datatype datatype,
 MPI_Aint * displacement );
-int   E_MPI_Type_size(
+double E_MPI_Type_size(
 MPI_Datatype datatype,
 int * size );
-int  E_MPI_Type_struct(
+double E_MPI_Type_struct(
 int count,
 int * blocklens,
 MPI_Aint * indices,
 MPI_Datatype * old_types,
 MPI_Datatype * newtype );
-int   E_MPI_Type_ub(
+double E_MPI_Type_ub(
 MPI_Datatype datatype,
 MPI_Aint * displacement );
-int  E_MPI_Type_vector(
+double E_MPI_Type_vector(
 int count,
 int blocklen,
 int stride,
 MPI_Datatype old_type,
 MPI_Datatype * newtype );
-int   E_MPI_Unpack(
+double E_MPI_Unpack(
 void * inbuf,
 int insize,
 int * position,
@@ -664,7 +664,7 @@ void * outbuf,
 int outcount,
 MPI_Datatype type,
 MPI_Comm comm );
-int   E_MPI_Pack(
+double E_MPI_Pack(
 void * inbuf,
 int incount,
 MPI_Datatype type,
@@ -672,7 +672,7 @@ void * outbuf,
 int outcount,
 int * position,
 MPI_Comm comm );
-int   E_MPI_Pack_size(
+double E_MPI_Pack_size(
 int incount,
 MPI_Datatype datatype,
 MPI_Comm comm,
